Disease_Detection_and_Classification_Using_Medical_Image_Data/requirements.txt
Werkzeug==0.15.3
numpy==1.16.1
Keras==2.2.4
tensorflow==1.15.2
gunicorn==19.9.0
Flask==1.0.4
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
pillow>=6.2.0

